+++ src/main/java/roomescape/payment/application/dto/PaymentResponse.java
@@ -1,9 +1,5 @@
 package roomescape.payment.application.dto;
 
-public record PaymentResponse(
-        String orderId,
-        String paymentKey,
-        Long totalAmount
-) {
+public interface PaymentResponse {
 
 }
