+++ src/main/java/roomescape/payment/infrastructure/dto/TossPaymentResponse.java
@@ -1,0 +1,10 @@
+package roomescape.payment.infrastructure.dto;
+
+import roomescape.payment.application.dto.PaymentResponse;
+
+public record TossPaymentResponse(
+        String orderId,
+        String paymentKey,
+        Long totalAmount
+) implements PaymentResponse {
+}
